backend/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Trading and market data
python-binance==1.0.19

# Machine Learning
stable-baselines3[extra]==2.2.1
gymnasium==0.29.1
torch>=1.13.0
tqdm==4.66.1
rich==13.7.0

# Data processing
pandas==2.1.3
numpy==1.24.3
ta==0.11.0

# Utilities
python-dotenv==1.0.0
websockets==12.0
redis==5.0.1
psutil==5.9.6

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2